requirements.txt
streamlit==1.30.0
pandas
pillow
requests
numpy
python-dotenv
psycopg2-binary
openai
pyngrok
tqdm
openpyxl
oracledb

Flask==3.0.3
psycopg2-binary==2.9.9
